moving print_helper.hpp to stansummary_helper.hpp
