+++ make/command
@@ -1,10 +1,9 @@
-.PRECIOUS: bin/print$(EXE)
-bin/print$(EXE) : bin/cmdstan/print.o
-	@mkdir -p $(dir $@)
-	$(LINK.c) -O$(O_STANC) $(OUTPUT_OPTION) $< $(LDLIBS)
-
 .PRECIOUS: bin/stanc$(EXE)
 bin/stanc$(EXE) : bin/cmdstan/stanc.o bin/libstanc.a
 	@mkdir -p $(dir $@)
 	$(LINK.c) -O$(O_STANC) $(OUTPUT_OPTION) $< $(LDLIBS_STANC)
 
+.PRECIOUS: bin/print$(EXE)
+bin/print$(EXE) : bin/cmdstan/print.o
+	@mkdir -p $(dir $@)
+	$(LINK.c) -O$(O_STANC) $(OUTPUT_OPTION) $< $(LDLIBS)
