+++ makefile
@@ -25,7 +25,7 @@ AR = ar
 # Library locations
 ##
 STAN ?= stan/
-MATH ?= $(STAN)lib/stan_math/
+MATH ?= $(STAN)lib/stan_math_2.8.0/
 -include $(MATH)make/libraries
 
 ##
