+++ src/cmdstan/print.cpp
@@ -3,7 +3,7 @@
 #include <iomanip>
 #include <ios>
 #include <stan/mcmc/chains.hpp>
-#include <cmdstan/print_helper.hpp>
+#include <cmdstan/stansummary_helper.hpp>
 #include <fstream>
 
 /**
